Breaking out plugins into own manager class.
