+++ dev/Bootstrapper.cs
@@ -35,81 +35,14 @@ namespace UltimaXNA
 
         private bool m_IsInitialized;
 
-        private List<IModule> m_Modules = new List<IModule>();
-
-        public Bootstrapper(string[] args)
-        {
-            GeneralExceptionHandler.Instance = new GeneralExceptionHandler();
-        }
-
-        private void LoadModule(IModule module)
-        {
-            m_Modules.Add(module);
-            module.Load();
-        }
-        
         public string BaseApplicationPath
         {
             get { return AppDomain.CurrentDomain.BaseDirectory; }
         }
 
-        private void ConfigureTraceListeners()
+        public Bootstrapper(string[] args)
         {
-            if (Debugger.IsAttached)
-            {
-                Tracer.RegisterListener(new DebugOutputEventListener());
-            }
-
-            Tracer.RegisterListener(new FileLogEventListener("debug.log"));
-
-            if (Settings.Debug.IsConsoleEnabled)
-            {
-                Tracer.RegisterListener(new ConsoleOutputEventListener());
-            }
-
-            Tracer.RegisterListener(new MsgBoxOnCriticalListener());
-        }
-
-        private void Configure()
-        {
-            ConfigurePlugins();
-        }
-
-        private void ConfigurePlugins()
-        {
-            DirectoryInfo directory = new DirectoryInfo(Path.Combine(BaseApplicationPath, "plugins"));
-
-            if (!directory.Exists)
-            {
-                return;
-            }
-
-            FileInfo[] assemblies = directory.GetFiles("*.dll");
-
-            foreach (FileInfo file in assemblies)
-            {
-                try
-                {
-                    Tracer.Info("Loading plugin {0}.", file.Name);
-
-                    Assembly assembly = Assembly.LoadFile(file.FullName);
-                    IEnumerable<Type> modules = assembly.GetTypes().Where(t => t.GetInterfaces().Contains(typeof (IModule)));
-
-                    foreach (Type module in modules)
-                    {
-                        Tracer.Info("Activating module {0}.", module.FullName);
-
-                        IModule instance = (IModule)Activator.CreateInstance(module);
-
-                        LoadModule(instance);
-                    }
-                }
-                catch (Exception e)
-                {
-                    Tracer.Warn("An error occurred while trying to load plugin. [{0}]", file.FullName);
-                    Tracer.Warn(e);
-                }
-            }
+            GeneralExceptionHandler.Instance = new GeneralExceptionHandler();
         }
 
         public void Initialize()
@@ -141,22 +74,37 @@ namespace UltimaXNA
             }
         }
 
-        private void Prepare()
-        {
-            AppDomain.CurrentDomain.UnhandledException += OnUnhandledException;
-            TaskScheduler.UnobservedTaskException += OnUnobservedTaskException;
-        }
-
         private void StartEngine()
         {
-            Prepare();
+            PrepareExceptionHandlers();
 
             using (UltimaGame engine = new UltimaGame())
             {
-                Configure();
-
                 engine.Run();
             }
+        }
+
+        private void ConfigureTraceListeners()
+        {
+            if (Debugger.IsAttached)
+            {
+                Tracer.RegisterListener(new DebugOutputEventListener());
+            }
+
+            Tracer.RegisterListener(new FileLogEventListener("debug.log"));
+
+            if (Settings.Debug.IsConsoleEnabled)
+            {
+                Tracer.RegisterListener(new ConsoleOutputEventListener());
+            }
+
+            Tracer.RegisterListener(new MsgBoxOnCriticalListener());
+        }
+
+        private void PrepareExceptionHandlers()
+        {
+            AppDomain.CurrentDomain.UnhandledException += OnUnhandledException;
+            TaskScheduler.UnobservedTaskException += OnUnobservedTaskException;
         }
 
         private void OnUnhandledException(object sender, UnhandledExceptionEventArgs e)
